requirements.txt
gradio
pandas
psycopg2-binary
selenium
beautifulsoup4
requests
python-dotenv
sqlalchemy
plotly